app/requirements.txt
PredictionIO==0.9.9
python-dotenv==0.8.2
